+++ frontend/src/hooks/hooks.ts
@@ -1,4 +1,4 @@
-import { AppDispatch, RootState } from "@/store.ts";
+import { AppDispatch, RootState } from "@/redux/store.ts";
 import { useDispatch, useSelector } from "react-redux";
 
 export const useAppDispatch = useDispatch.withTypes<AppDispatch>();
