+++ frontend/src/hooks/use-window-resize-delta.tsx
