+++ frontend/src/main.tsx
@@ -1,11 +1,11 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import Login from "@/Login.tsx";
+import Login from "@/Login/Login.tsx";
 import { Toaster } from "@/components/ui/toaster.tsx";
 import { ToastProvider } from "@/components/ui/toast.tsx";
 import { Provider } from "react-redux";
-import store from "./store";
+import store from "./redux/store.ts";
 import { BrowserRouter } from "react-router";
 
 createRoot(document.getElementById("root")!).render(
