+++ frontend/src/scene/animations/Animator.tsx
@@ -10,30 +10,37 @@ import { Object3D, Quaternion, Vector3 } from "three";
 import { getAnimation } from "@/util/LimitUtils.ts";
 
 function Animator() {
+  // Access the Three.js state, including the scene and camera
   const state = useThree();
 
+  // Retrieve the list of model parts from the Redux store
   const modelParts = useAppSelector((state) => state.modelParts.partIds);
 
+  // Ref to store animations for objects, mapping each object to its target position or rotation
   const animationsRef = useRef<
     Map<Object3D, { target: Vector3 | Quaternion; topic: string }>
   >(new Map());
 
+  // Update animations on every frame
   useFrame((_rootState, delta) => {
     const damping = 1;
-    // For a smooth animation the lerp factor is based on the delta including a damping factor
+    // Calculate the lerp factor for smooth animations
     const lerpFactor = 1 - Math.exp(-damping * delta);
     animationsRef.current.forEach((data, object) => {
       if (data.target instanceof Vector3) {
+        // Smoothly interpolate the object's position
         object.position.lerp(data.target, lerpFactor);
 
-        // If the object is close we end the animation by setting the position directly
+        // Stop the animation if the object is close to the target
         if (object.position.distanceTo(data.target) < 0.0001) {
           object.position.copy(data.target);
           animationsRef.current.delete(object);
         }
       } else {
+        // Smoothly interpolate the object's rotation
         object.quaternion.slerp(data.target, lerpFactor);
 
+        // Stop the animation if the object is close to the target rotation
         if (object.quaternion.angleTo(data.target) < 0.0001) {
           object.quaternion.copy(data.target);
         }
@@ -41,23 +48,27 @@ function Animator() {
     });
   });
 
+  // Callback to handle incoming animation messages
   const animationCallback = useCallback(
     (msg: RelativePositionEventMessage) => {
       const name = msg.message.topic.split(".").reverse()[0];
 
       if (modelParts.length == 0) {
-        // Not initialized yet
+        // Skip processing if the model parts are not initialized
         return;
       }
 
+      // Find the object in the scene by its name
       const selectedObject = state.scene.getObjectByName(name);
 
       if (selectedObject) {
+        // Calculate the target position or rotation for the animation
         const localTarget = getAnimation(
           selectedObject,
           state.scene,
           msg.message.percentage
         );
+        // Store the animation data in the ref
         animationsRef.current.set(selectedObject, {
           target: localTarget,
           topic: msg.message.topic,
@@ -71,6 +82,7 @@ function Animator() {
     [state.scene, modelParts]
   );
 
+  // Memoize the list of topics to subscribe to for animations
   const topics = useMemo(
     () =>
       modelParts.map((modelPart) => {
@@ -79,13 +91,14 @@ function Animator() {
     [modelParts]
   );
 
+  // Subscribe to filtered WebSocket messages for animation updates
   useFilteredWebsocket(
     topics,
     MessageType.ANIMATION_RELATIVE,
     animationCallback
   );
 
-  return null;
+  return null; // This component does not render anything
 }
 
 export default Animator;
