+++ frontend/src/scene/animations/debug/Publisher.tsx
