+++ frontend/src/sidebar/replay/ReplayGroup.tsx
@@ -6,7 +6,7 @@ import {
 } from "@/components/ui/sidebar.tsx";
 import { StartReplay } from "@/sidebar/replay/StartReplay.tsx";
 import { useAppSelector } from "@/hooks/hooks.ts";
-import { selectModelParts } from "@/slices/ModelPartSlice.ts";
+import { selectModelParts } from "@/redux/slices/ModelPartSlice.ts";
 
 export function ReplayGroup() {
   const modelParts = useAppSelector(selectModelParts);
